+++ R/import_helper.r
@@ -1,5 +1,5 @@
 #This internal helper function prints basic information about a dataset and is used for import function
-import.info <- function(tmp, device, tz, Id.colname) {
+import.info <- function(tmp, device, tz, Id.colname, dst_adjustment) {
   #give info about the file
   min.time <- min(tmp$Datetime)
   max.time <- max(tmp$Datetime)
@@ -13,18 +13,47 @@ import.info <- function(tmp, device, tz, Id.colname) {
     dplyr::mutate(pct = (n/sum(n)) %>% scales::percent(),
                   interval.time = interval.time %>% lubridate::as.duration())
 
+  #number of Ids
+  n_ids <- tmp %>% dplyr::group_keys() %>% nrow()
+  #number of files
+  n_files <- tmp$file.name %>% unique() %>% length()
+  
+  #check for dst_adjustment
+  dst_info <- 
+    tmp %>% dplyr::group_by(file.name, .add = TRUE) %>% dst_change_summary()
+  
+  #prepare dst info
+  if(nrow(dst_info) != 0) {
+    dst_info <- 
+      paste0(
+        "Observations in the following ", 
+        dst_info$file.name %>% unique() %>% length(),
+        " file(s) cross to or from daylight savings time (DST): \n",
+        dst_info$file.name %>% unique() %>% paste0(collapse = "\n"), "\n")
+    if(dst_adjustment) {
+      dst_info <- paste0(dst_info, "The Datetime column was adjusted in these files. For more info on what that entails see `?dst_change_handler`.\n")
+    } else {
+      dst_info <- paste0(dst_info, "Please make sure that the timestamps in the source files correctly reflect these changes from DST<>ST. \nTo adjust datetimes after a jump, set `dst_adjustment = TRUE` or see `?dst_change_handler` for manual adjustment.\n")
+    }
+  } else {
+    dst_info <- NULL
+  }
+  
+  #print all infos
     cat(
-    "Successfully read in ", nrow(tmp), " observations from ", device, "-file", 
     "\n",
+    "Successfully read in ", format(nrow(tmp), big.mark = "'"), 
+    " observations across ", n_ids, " Ids from ",  n_files, " ", device, "-file(s).\n",
     "Timezone set is ", tz, ".\n", 
     if(lubridate::tz(tmp$Datetime) != Sys.timezone()) {
       paste0(
         "The system timezone is ",
         Sys.timezone(),
         ". Please correct if necessary!\n")},
-    "Start: ", format(min.time), "\n",
-    "End: ", format(max.time), "\n",
-    "Timespan: " , diff(c(min.time, max.time)) %>% format(digits = 2), "\n",
+    dst_info, "\n",
+    "First Observation: ", format(min.time), "\n",
+    "Last Observation: ", format(max.time), "\n",
+    "Timespan: " , diff(c(min.time, max.time)) %>% format(digits = 2), "\n\n",
     "Observation intervals: \n",
     sep = "")
   utils::capture.output(interval.time)[c(-1,-2,-4)] %>% cat(sep = "\n")